Make Revelio an importable library
