+++ src/main/webapp/library.js
@@ -1,0 +1,21 @@
+import Sources from './sources'
+import Workspaces, { Workspace } from './workspaces/workspaces'
+import SimpleSearch from './simple-search'
+import Search from './search/search'
+import SearchResults from './search/results'
+import SearchDetails from './search/details'
+import ResultForms from './result-forms'
+import SearchForms from './search-forms'
+import About from './about'
+module.exports = {
+  About,
+  ResultForms,
+  Search,
+  SearchDetails,
+  SearchForms,
+  SearchResults,
+  SimpleSearch,
+  Sources,
+  Workspace,
+  Workspaces,
+}
